requirements.txt
alembic==1.7.5
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
generate-password==1.0.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
permission==0.4.1
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.8
Werkzeug==2.0.2
zipp==3.7.0
